Correcting draw rates (2nd pool) for SR items
When creating banner data for this banner, I did not generate the complete banner data (I forgot to generate draw rates from the 2nd pool) before the banner ended. Because of this, I had to calculate draw rates for the 2nd pool based on rates from the first pool myself. There were 2 possible methods of calculation from what I knew at the time, and I uploaded the banner with with the first method, which I thought was more likely. Now that more data has come in, I now know that it is the 2nd method that is correct, so I am rectifying this mistake.